b2b unit added

+++ skillrat/usermanagement/src/main/java/com/skillrat/usermanagement/dto/AddressDTO.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -16,8 +18,8 @@ public class AddressDTO {
     private String city;
     private String state;
     private String country;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private String postalCode;
     private Long userId;
     private Boolean isDefault;
